add search complete and update tests
