+++ pkg/resolver/mocks/mock.json
@@ -4,7 +4,18 @@
     {
         "uid":"local-cluster/e12c2ddd-4ac5-499d-b0e0-20242f508afd",
         "resourceString":"templates",
-        "properties": {"kind":"Template"}
+        "properties":{
+            "_clusterNamespace":"local-cluster-ns",
+            "_hubClusterResource":true,
+            "apigroup":"template.openshift.io",
+            "apiversion":"v1",
+            "created":"2021-07-14T10:20:37Z",
+            "kind":"Template",
+            "label":{"samples.operator.openshift.io/managed":"true"},
+            "name":"eap-cd-starter-s2i",
+            "namespace":"openshift"
+        },
+        "Metadata":{"OwnerUID":""}
     },
     {
         "uid":"local-cluster/13250bc4-865c-41db-a8f2-05bec0bd042b",
@@ -38,4 +49,6 @@
         "DestUID": "local-cluster/411e30e4-f773-41a6-b745-24c93c173f45",
         "SourceKind": "ReplicaSet",
         "DestKind": "Deployment"
+    }
+    ]
 }
